+++ src/features/api/index.ts
@@ -4,6 +4,7 @@ import { settings } from "../../constants/settings";
 import { settingsStore } from "../../scripts/settings";
 import { addCurrentInfo } from "./features/current";
 import { addPlaybackControl } from "./features/player";
+import { addSettingsAPI } from "./features/settings/settings";
 import { addLegacyApi } from "./legacy";
 
 /**
@@ -11,12 +12,14 @@ import { addLegacyApi } from "./legacy";
  */
 export const startApi = (mainWindow: BrowserWindow) => {
   const expressApp = express();
+  expressApp.use(express.json());
   expressApp.get("/", (req, res) => res.send("Hello World!"));
 
   // add features
   addLegacyApi(expressApp, mainWindow);
   addPlaybackControl(expressApp, mainWindow);
   addCurrentInfo(expressApp);
+  addSettingsAPI(expressApp, mainWindow);
 
   const port = settingsStore.get<string, number>(settings.apiSettings.port);
   const expressInstance = expressApp.listen(port, "127.0.0.1");