+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [next]
 
 - Fixed [#403](https://github.com/Mastermindzh/tidal-hifi/issues/403) "cannot read shuffle of undefined" error
+- Added an API to add & delete entries from the skippedArtists list in the settings. fixes [#405](https://github.com/Mastermindzh/tidal-hifi/issues/405)
+  - `GET /settings/skipped-artists` -> get list of skipped artists
+  - `POST /settings/skipped-artists` -> add to the list of skipped artists
+  - `POST /settings/skipped-artists/delete` -> delete from the list of skipped artists
+  - `POST /settings/skipped-artists/current` -> skip the current artist
+  - `DELETE /settings/skipped-artists/current` -> delete the current artist from the skip list
 
 ## [5.12.0]